Routing with base path
- Support apps deployed from a sub-path instead of the domain root
