+++ src/CmsConfig.ts
@@ -1,15 +1,4 @@
-import { IPageConfigNode } from "./PageStructureReader";
-
-export interface IConfigData {
-    /** Where to load plugins from */
-    pluginsBaseUrl: string;
-    /** Map of pluginUri => config object */
-    plugins: Record<string, unknown>;
-    /** A tree structure describing which pages are available, with child modules */
-    pages: IPageConfigNode[];
-    /** Modules that are rendered directly in the root component (passed as Cms children prop) */
-    rootModules: Record<string, IPageConfigNode[]>;
-}
+import { IConfigData } from "./IConfigData";
 
 export class CmsConfig {
     private data: IConfigData;
